+++ src/combinator.rs
@@ -221,7 +221,7 @@ impl<T: Action> Action for Combinator<T> {
   fn exec<'text>(
     &self,
     input: &mut Input<'text, &mut Self::State, &mut Self::Heap>,
-  ) -> Option<Output<'text, T::Value>> {
+  ) -> Option<Output<T::Value>> {
     self.action.exec(input)
   }
 }