+++ src/combinator/provided/till.rs
@@ -12,18 +12,12 @@ use crate::{
 pub trait Till<State, Heap> {
   /// Check if the rest of input text contains this instance.
   /// Return the output after digesting the instance if found.
-  fn exec<'text>(
-    &self,
-    input: &mut Input<'text, &mut State, &mut Heap>,
-  ) -> Option<Output<'text, ()>>;
+  fn exec<'text>(&self, input: &mut Input<'text, &mut State, &mut Heap>) -> Option<Output<()>>;
 }
 
 impl<State, Heap> Till<State, Heap> for &str {
   #[inline]
-  fn exec<'text>(
-    &self,
-    input: &mut Input<'text, &mut State, &mut Heap>,
-  ) -> Option<Output<'text, ()>> {
+  fn exec<'text>(&self, input: &mut Input<'text, &mut State, &mut Heap>) -> Option<Output<()>> {
     input
       .rest()
       .find(self)
@@ -33,20 +27,14 @@ impl<State, Heap> Till<State, Heap> for &str {
 
 impl<State, Heap> Till<State, Heap> for String {
   #[inline]
-  fn exec<'text>(
-    &self,
-    input: &mut Input<'text, &mut State, &mut Heap>,
-  ) -> Option<Output<'text, ()>> {
+  fn exec<'text>(&self, input: &mut Input<'text, &mut State, &mut Heap>) -> Option<Output<()>> {
     self.as_str().exec(input)
   }
 }
 
 impl<State, Heap> Till<State, Heap> for char {
   #[inline]
-  fn exec<'text>(
-    &self,
-    input: &mut Input<'text, &mut State, &mut Heap>,
-  ) -> Option<Output<'text, ()>> {
+  fn exec<'text>(&self, input: &mut Input<'text, &mut State, &mut Heap>) -> Option<Output<()>> {
     input
       .rest()
       .find(*self)
@@ -56,11 +44,8 @@ impl<State, Heap> Till<State, Heap> for char {
 
 impl<State, Heap> Till<State, Heap> for () {
   #[inline]
-  fn exec<'text>(&self, _: &mut Input<'text, &mut State, &mut Heap>) -> Option<Output<'text, ()>> {
-    Some(Output {
-      value: (),
-      rest: "",
-    })
+  fn exec<'text>(&self, input: &mut Input<'text, &mut State, &mut Heap>) -> Option<Output<()>> {
+    unsafe { input.digest_unchecked(input.rest().len()) }.into()
   }
 }
 
@@ -101,28 +86,28 @@ mod tests {
       till("end".to_string()).exec(&mut Input::new("123end456", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
-        rest: "456"
+        digested: 6
       })
     );
     assert_eq!(
       till("end").exec(&mut Input::new("123end456", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
-        rest: "456"
+        digested: 6
       })
     );
     assert_eq!(
       till(';').exec(&mut Input::new("123;456", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
-        rest: "456"
+        digested: 4
       })
     );
     assert_eq!(
       till(()).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
-        rest: ""
+        digested: 3
       })
     );
   }